CE-2487 add diff link
